requirements.txt
certifi==2019.6.16
chardet==3.0.4
idna==2.8
oauthlib==3.1.0
PySocks==1.7.0
PyYAML==5.1.2
requests==2.22.0
requests-oauthlib==1.2.0
six==1.12.0
tweepy==3.8.0
urllib3==1.25.3
beautifulsoup4
pyrebase